+++ src/types/index.ts
@@ -1,9 +1,14 @@
+export type ValidatedPublication = {
+  classes: ClassCollection[];
+  maturity: string;
+};
+
 export type ClassCollection = {
-  classURI: string,
-  className: string, 
-  count: number,
-  objects: ValidatedSubject[]
-}
+  classURI: string;
+  className: string;
+  count: number;
+  objects: ValidatedSubject[];
+};
 
 export type ParsedSubject = {
   uri: string;
@@ -11,12 +16,10 @@ export type ParsedSubject = {
   properties: ParsedProperty[];
 };
 
-
 export type ParsedProperty = {
   path: string;
   value: string | ParsedSubject;
 };
-
 
 export type ValidatedSubject = {
   uri: string;
@@ -29,29 +32,29 @@ export type ValidatedSubject = {
   properties: ValidatedProperty[] | ProcessedProperty[];
 };
 
-
 export type ValidatedProperty = {
   name: string;
   targetClass?: string;
   description: string;
   path: string;
   value: string[] | ValidatedSubject[];
-  minCount?: number
+  minCount: number;
   maxCount?: number;
   actualCount: number;
+  maturityLevel?: string;
   valid: boolean;
-  example?: string
+  example?: string;
 };
 
 export type DocumentType = {
   id?: string;
-  label: string
-}
+  label: string;
+};
 
 export type ProcessedProperty = {
   name: string;
   path: string;
   value: string[] | ValidatedSubject[];
   actualCount: number;
-  example?: string
+  example?: string;
 };
